fix(gitlab): paginate MR diffs fetch
GitLab's /diffs endpoint paginates at 20 files per page. Without
--paginate, fetchMrDiffs only sees page 1, causing draft-note.ts batch
to fail validation with "File X not found in MR diff" for any file past
the first page (hit on MR !426 with 30+ files).

Move parseGlabPaginated from discussions.ts to diff.ts (the lower-level
module) and use --paginate + parseGlabPaginated in fetchMrDiffs. Tests
move alongside. discussions.ts re-exports parseGlabPaginated to keep
the external import surface stable.
