+++ plugins/gitlab/skills/merge-request/scripts/diff.test.ts
@@ -1,5 +1,38 @@
 import { describe, expect, it } from "bun:test";
-import { buildPosition, isLineInDiff, parseDiffHunks, validateLineInDiff } from "./diff";
+import {
+  buildPosition,
+  isLineInDiff,
+  parseDiffHunks,
+  parseGlabPaginated,
+  validateLineInDiff,
+} from "./diff";
+
+describe("parseGlabPaginated", () => {
+  it("parses a single page", () => {
+    const result = parseGlabPaginated('[{"id": 1}]');
+    expect(result).toEqual([{ id: 1 }]);
+  });
+
+  it("fixes concatenated pages", () => {
+    const result = parseGlabPaginated('[{"id": 1}][{"id": 2}]');
+    expect(result).toEqual([{ id: 1 }, { id: 2 }]);
+  });
+
+  it("handles whitespace between pages", () => {
+    const result = parseGlabPaginated('[{"id": 1}]\n[{"id": 2}]');
+    expect(result).toEqual([{ id: 1 }, { id: 2 }]);
+  });
+
+  it("handles three concatenated pages", () => {
+    const result = parseGlabPaginated('[{"a":1}][{"b":2}][{"c":3}]');
+    expect(result).toEqual([{ a: 1 }, { b: 2 }, { c: 3 }]);
+  });
+
+  it("preserves arrays inside objects", () => {
+    const result = parseGlabPaginated('[{"notes": [1, 2]}]');
+    expect(result).toEqual([{ notes: [1, 2] }]);
+  });
+});
 
 describe("parseDiffHunks", () => {
   it("parses a single hunk", () => {
